more bytecode

+++ compiler/compiler/src/main/scala/io/github/spaceman1701/oxidizer/compiler/bytecode/Bytecode.scala
@@ -5,11 +5,11 @@ sealed trait Instruction {
   /*
   Basic Load/Store instructions. For simplicity only wide insturctions.
    */
-  case class Store(loc: Long) extends Instruction //objref | value -> [] (convert to objref and place on data stack)
+  case class Store(loc: Int) extends Instruction //objref | value -> [] (convert to objref and place on data stack)
   case class LoadConstInt(value: Long) extends Instruction
   case class LoadConstFloat(value: Double) extends Instruction
   case class LoadConstStr(ptr: Long) extends Instruction //load a string from the constant string areas
-  case class LoadVar(loc: Long) extends Instruction
+  case class LoadVar(loc: Int) extends Instruction
 
   /*
   Native list slice. Allows complex slices to be done in one instruction
